requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
pillow==10.0.0
trimesh==3.22.1
open3d==0.17.0
flask==2.3.2
flask-socketio==5.3.4
moderngl==5.8.2
pygame==2.5.0
transformers==4.30.2
torch==2.0.1
scikit-image==0.21.0
spacy==3.6.1
nltk==3.8.1
sqlalchemy==2.0.19
redis==4.6.0
celery==5.3.1
psutil==5.9.5
matplotlib==3.7.2
scipy==1.11.1
sklearn==1.3.0
requests==2.31.0
websockets==11.0.3
pyopengl==3.1.7
pyopengl-accelerate==3.1.7
noise==1.2.2
perlin-noise==1.12
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
aiofiles==23.1.0
aioredis==2.0.1
huggingface-hub==0.16.4
torchvision==0.15.2
torchaudio==0.15.2
onnx==1.14.1
onnxruntime==1.15.1
mysql-connector-python==8.0.33